Revert "Feature/visual taking damage"
