--- Astral Drift/Assets/GameObjects/Enemies/Building blocks/scripts/General/FlashOnHit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b35883d0e7c6a6140a79857fd7560e06
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
