+++ Astral Drift/Assets/GameObjects/Player/Prefabs/Player.prefab
@@ -151,21 +151,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a632c84da2ede0046a2e9ffc3014fccb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  flashOnHit:
-    m_PersistentCalls:
-      m_Calls:
-      - m_Target: {fileID: 7313411074426475004}
-        m_TargetAssemblyTypeName: FlashOnHit, Assembly-CSharp
-        m_MethodName: Flash
-        m_Mode: 1
-        m_Arguments:
-          m_ObjectArgument: {fileID: 0}
-          m_ObjectArgumentAssemblyTypeName: UnityEngine.Object, UnityEngine
-          m_IntArgument: 0
-          m_FloatArgument: 0
-          m_StringArgument: 
-          m_BoolArgument: 0
-        m_CallState: 2
   maxHitpoints: 100
   collisionDamageToPlayer: 0
   collisionDamageToEnemy: 60
@@ -534,25 +519,6 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 7064678698491398669, guid: 42bc8c314e6102e499c4848dec31a5ec, type: 3}
   m_PrefabInstance: {fileID: 6231379382846914988}
   m_PrefabAsset: {fileID: 0}
---- !u!1 &4556334389132366107 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 7584369034610151607, guid: 42bc8c314e6102e499c4848dec31a5ec, type: 3}
-  m_PrefabInstance: {fileID: 6231379382846914988}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &7313411074426475004
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 4556334389132366107}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b35883d0e7c6a6140a79857fd7560e06, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  flashTexture: {fileID: 2800000, guid: ee24da3c9c95dbe42b5dc700c0d6fb04, type: 3}
-  flashTime: 0.05
 --- !u!1001 &6516377511938588886
 PrefabInstance:
   m_ObjectHideFlags: 0
