+++ Astral Drift/Assets/UIObjects/Prefabs/Gameplay UI.prefab
@@ -734,108 +734,6 @@ MonoBehaviour:
   m_hasFontAssetChanged: 0
   m_baseMaterial: {fileID: 0}
   m_maskOffset: {x: 0, y: 0, z: 0, w: 0}
---- !u!1001 &3246268723145064087
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 1548219102140828378}
-    m_Modifications:
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_Pivot.x
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_Pivot.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_RootOrder
-      value: 5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_AnchorMax.x
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_AnchorMax.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_AnchorMin.x
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_AnchorMin.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_SizeDelta.x
-      value: 100
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_SizeDelta.y
-      value: 100
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_AnchoredPosition.x
-      value: -10
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_AnchoredPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 3246268723792583043, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-      propertyPath: m_Name
-      value: HealthBar
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
---- !u!224 &664512277 stripped
-RectTransform:
-  m_CorrespondingSourceObject: {fileID: 3246268723792583042, guid: 97a1daca47b9b0f48b8c4e75c01ad35f, type: 3}
-  m_PrefabInstance: {fileID: 3246268723145064087}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &1794395022433334901
 GameObject:
   m_ObjectHideFlags: 0
