--- Astral Drift/Assets/UIObjects/Prefabs/HealthBar.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 4dfffbc696f96f347b6694743b11ecef
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
