--- Astral Drift/Assets/UIObjects/Prefabs/HealthBarUI.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 97a1daca47b9b0f48b8c4e75c01ad35f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
