--- Astral Drift/Assets/UIObjects/Scripts/HealthBarUIEnemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfbcd7905c36f184b8994931f853caac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 380
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
