+++ Astral Drift/Assets/_GeneralAssets/Scenes/A-B Testing/ShowCase Test A.unity
@@ -4288,10 +4288,6 @@ PrefabInstance:
       propertyPath: shootingOffsetTimer
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 8272062397749759515, guid: 07b25bd21d7b96248ab88bfdf04edbc2, type: 3}
-      propertyPath: moveClockwise
-      value: 0
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 07b25bd21d7b96248ab88bfdf04edbc2, type: 3}
 --- !u!1001 &1495927569
@@ -4706,10 +4702,6 @@ PrefabInstance:
     - target: {fileID: 5799372889425806587, guid: 07b25bd21d7b96248ab88bfdf04edbc2, type: 3}
       propertyPath: shootingOffsetTimer
       value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 8272062397749759515, guid: 07b25bd21d7b96248ab88bfdf04edbc2, type: 3}
-      propertyPath: moveClockwise
-      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 07b25bd21d7b96248ab88bfdf04edbc2, type: 3}
@@ -5481,9 +5473,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e79b91ab965c46f479854171f2491304, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  flashOnHit:
-    m_PersistentCalls:
-      m_Calls: []
   maxHitpoints: 1400
 --- !u!50 &1961153588
 Rigidbody2D:
@@ -7116,9 +7105,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e79b91ab965c46f479854171f2491304, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  flashOnHit:
-    m_PersistentCalls:
-      m_Calls: []
   maxHitpoints: 1400
 --- !u!1001 &6233368479132000416
 PrefabInstance:
@@ -7145,11 +7131,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 69877416, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_SizeDelta.x
-      value: 37
+      value: 23
       objectReference: {fileID: 0}
     - target: {fileID: 69877416, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_SizeDelta.y
-      value: 266
+      value: 686
       objectReference: {fileID: 0}
     - target: {fileID: 69877416, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_AnchoredPosition.x
@@ -7217,11 +7203,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1612824967, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_SizeDelta.x
-      value: 21
+      value: 23
       objectReference: {fileID: 0}
     - target: {fileID: 1612824967, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_SizeDelta.y
-      value: 250
+      value: 686
       objectReference: {fileID: 0}
     - target: {fileID: 1612824967, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_AnchoredPosition.x
@@ -7245,11 +7231,11 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1833960352, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_SizeDelta.x
-      value: 21
+      value: 23
       objectReference: {fileID: 0}
     - target: {fileID: 1833960352, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_SizeDelta.y
-      value: 250
+      value: 686
       objectReference: {fileID: 0}
     - target: {fileID: 1833960352, guid: 69a9088e927d2d54c864898143f10829, type: 3}
       propertyPath: m_AnchoredPosition.x
