+++ Astral Drift/ProjectSettings/EditorBuildSettings.asset
@@ -17,10 +17,7 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/_GeneralAssets/Scenes/Release v0.2.unity
     guid: 621d66e8568333049918d64249c3c652
-  - enabled: 0
+  - enabled: 1
     path: Assets/_GeneralAssets/Scenes/Showcases/Player.unity
     guid: 4e084f03488b7114292db8e2cdc90d0c
-  - enabled: 1
-    path: Assets/_GeneralAssets/Scenes/Showcases/AllEnemies.unity
-    guid: 59f18329d864df54fae2f7c890a2bede
   m_configObjects: {}
